Extract incoming damage projector
